fix experimentAttribute seeds, split seeds into diffrent files
